feat: implement season management and user game statistics
- Added SeasonId as a new identifier type for seasons.
- Introduced user game statistics tracking with the UserGameStats struct.
- Implemented CRUD operations for seasons in the PgSeasonRepo.
- Enhanced leaderboard functionality to aggregate user statistics by season and game.
- Updated routes for season management and leaderboard queries.
- Created database migrations for seasons and user game statistics tables.

+++ crates/sw-domain/src/lobby.rs
@@ -130,7 +130,7 @@ pub struct PlayerState {
     pub state: JoinRequestState,
     pub rank: Option<usize>,
     pub prize: Option<f64>,
-    pub wars_point: Option<f64>,
+    pub wars_point: Option<i64>,
     pub claim_state: Option<ClaimState>,
     pub last_ping: Option<u64>,
     pub joined_at: i64,
